Adding the content model to the leadspace block (#831)
* Adding the content model to the leadspace block

I added the content model to the leadspace block. I utilized the same format as was used for Leadspace. I also added an anchor link at the top for the user to quickly get to this section.

* Added links to the content model

I added links to the components in the content model so they take the user to the appropriate component pages. I also added in a link that takes the user to the git wiki which has a list of all character counts.

* Changing table links and labels

- Changed the link for CTA (button) to go to "CTA" component
- Changed the label from "Video" to "Video Player"

Co-authored-by: JennySanchez <45213839+JennySanchez@users.noreply.github.com>