Content group simple - content model
This is a new variation of content model that i'm trying out for the content group simple. The reason it is different is because there is no page for content item which is used in this component. So either a recommend this content model layout or some variation of it or we could add a new page for content item. We currently have one for content item-horizontal, but not just content item. Please give thoughts.