Added content model to video player (#838)
* Added content model to video player

Added content model to video player and also an anchor link to get to the content model

* Changing 'showCaption' to 'hideCaption'

As @photodow pointed out, this option should be 'showCaption' instead of 'hideCaption'. It is recommended to change the Storybook documentation to match this.

* Adding security to external links

There are links to the character count standards and to the box design and functional specs. I've added rel="noreferrer noopener nofollow" to both to add security and follow SEO best practices.

* Removed fields on request

Removed the Caption and "hideCaption" fields. Also, changed autoplay and aspect ratio "required" fields to "No"

Co-authored-by: kodiakhq[bot] <49736102+kodiakhq[bot]@users.noreply.github.com>