Added content model to image with caption (#839)
* Added content model to image with caption

Added content model to image with caption

* Updates

Added rel="noopener noreferrer nofollow" to some of the links for better security and SEO.
Added a notes column to the content model so there can be more information specific to the content fields.

* Changing "rel" attributes

Removed nofollow and noopener from "rel" in both external links.

* Adding line in table header

Co-authored-by: JennySanchez <45213839+JennySanchez@users.noreply.github.com>
Co-authored-by: kodiakhq[bot] <49736102+kodiakhq[bot]@users.noreply.github.com>