Content block segmented content model (#847)
* Content block segmented content model

This follows the new content model layout as the 'content-group-simple' because the content item component page does not exist. I've added an anchor link to this page and added security to the outgoing Box link in the design and functional specifications section.

* Changing content group to content item

Update content model from James' feedback to switch content item to content group.

* Adding line breaks

* Removing content group simple fields

* Update CTA (card) link to be anchor

Co-authored-by: JennySanchez <45213839+JennySanchez@users.noreply.github.com>
Co-authored-by: kodiakhq[bot] <49736102+kodiakhq[bot]@users.noreply.github.com>