Content block with cards content model (#849)
* Content block with cards content model

I've added the content model to the content block with cards as well as an anchor link to get to content guidance. I also added extra security to the Box link that already exists on this page.

* Changing card cardinality

Changed the cardinality of the card from "2-9" to "2-6" based on previous documentation.

* Updated with card group info

Updated the content model to just include card group as that is the dependency. Also add some notes to CTA (card) in the content model and got rid of the some of the "rel" attributes in the external links.

* Removing TOC headline override

* Updating CTA (card) to include anchor link

Co-authored-by: JennySanchez <45213839+JennySanchez@users.noreply.github.com>
Co-authored-by: kodiakhq[bot] <49736102+kodiakhq[bot]@users.noreply.github.com>