Adding the content model to callout quote (#833)
* Adding the content model to callout quote

I've added an anchor link for content guidance and the content model to the page.

* Fixed the line breaks

There was an issue with the line breaks which is why it didn't deploy correctly. This will fix the issue.

* Adding security to external links

There are links to the character count standards and to the box design and functional specs. I've added rel="noreferrer noopener nofollow" to both to add security and follow SEO best practices.

* Update src/pages/components/callout-quote.mdx

This adds in code examples to the quotation mark type field.

Co-authored-by: James Dow <photodow@gmail.com>

* Adding the corner bracket option

* Making quotation mark type not required

This is not required since it has a default value.

* Update src/pages/components/callout-quote.mdx

Co-authored-by: James Dow <photodow@gmail.com>

* Update src/pages/components/callout-quote.mdx

Co-authored-by: James Dow <photodow@gmail.com>

* Updating link breaks

The previous line breaks were broken so these should fix them.

* Updating the example quotations

* Updating CTA (link) to CTA (text)

Co-authored-by: James Dow <photodow@gmail.com>
Co-authored-by: JennySanchez <45213839+JennySanchez@users.noreply.github.com>