Button group content model (#842)
* Button group content model

I added the very simple content model to button group. I also added an anchor link to get to the content model.

* Updating link to Carbon button

The link was going to the CTA (button) of the "Carbon for IBM" website, but now we switched to the Carbon Button since that is the true dependency.

* Add rel attributes and target="_blank"

Co-authored-by: JennySanchez <45213839+JennySanchez@users.noreply.github.com>
Co-authored-by: James Dow <photodow@gmail.com>
Co-authored-by: kodiakhq[bot] <49736102+kodiakhq[bot]@users.noreply.github.com>