Update opening-link.mdx
Update the content format and take the Mobile and Tablet information (listed below for record purpose) off based on the feedback received from Mike and the Carbon Team.

---

<Title>When users relying on the mobile device to browse the information</Title>

New windows on the mobile often cause disorientation, with users not realizing that a new window or tab has opened. This problem is exacerbated on mobile, where the old window is never visible.

<Title>When users relying on the tablet device to browse the information</Title>

New Windows devices will require users to have multiple windows and tabs for the browser, so the browsing experience is not easy to manage.