docs(developing): adding framework documentation for developing section (#1061)
* docs(developing): adding framework documentation for developing section

* Update react.mdx

This is wonderful. It sounds so easy it makes me think I can be a developer! :) 

Just some capitalization edits here to bring in line with IBM Style.

* Update web-components.mdx

Just some capitalization tweaks. @jeffchew I would also remove the parentheses in both of the install H2 headings. I left them there, just in case there was a reason for them to be there.

* Update index.mdx

Some capitalization tweaks :)

* Update index.mdx

@jeffchew I changed the H2 heading name to be different from the page title, but you should take a look to make sure it's okay.

* Update src/data/nav-items.yaml

Co-authored-by: Jan Child <44503588+janchild@users.noreply.github.com>

* Update src/data/nav-items.yaml

Co-authored-by: Jan Child <44503588+janchild@users.noreply.github.com>

* Update src/pages/developing/index.mdx

Co-authored-by: Jan Child <44503588+janchild@users.noreply.github.com>

* docs(developing): removing parenthesis in the H2's

Co-authored-by: Jan Child <44503588+janchild@users.noreply.github.com>